pyproject.toml
[project]
name = "social-assistant"
version = "0.1.0"
description = "ARTOGO social media automation assistant for Threads"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "slack-bolt>=1.20",
    "anthropic>=0.40",
    "openai>=1.50",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
